+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 -----
 - Add ``template_amplicon_report.tex`` to ``resolwebio/latex`` Docker image
   assets
+- Add SnpEff tool and bokeh assets to ``resolwebio/dnaseq`` Docker image
 
 Fixed
 -----